server/requirements.txt
Flask>=2.2.0
Flask-WTF>=1.1.1
Flask-SQLAlchemy>=3.0.0
WTForms>=3.0.0
python-dotenv>=1.0.0